requirements.txt
Flask>=2.2
Pillow>=9.0
numpy>=1.24
